refactored image detection and moce process into separate files
