--- idling.py
@@ -1,3 +1,0 @@
-"""
-TODO: idling
-"""
