+++ openai-image-detection.py
@@ -1,0 +1,72 @@
+import os
+from dotenv import load_dotenv
+from openai import OpenAI
+import base64
+import json
+import requests
+
+load_dotenv()
+
+client = OpenAI(
+    # This is the default and can be omitted
+    api_key=os.environ.get("OPENAI_API_KEY"),
+)
+
+"""
+Identifies the object in an image using the OpenAI API's image recognition model.
+
+Parameters:
+- image_path (str): The file path of the image to be analyzed.
+
+Returns:
+- dict: A dictionary containing the identified object and its characteristics.
+"""
+def identify_image(image_path):
+    api_key = os.environ.get("OPENAI_API_KEY")
+
+    # Function to encode the image
+    def encode_image(image_path):
+        with open(image_path, "rb") as image_file:
+            return base64.b64encode(image_file.read()).decode('utf-8')
+
+    # Getting the base64 string
+    base64_image = encode_image(image_path)
+
+    headers = {
+        "Content-Type": "application/json",
+        "Authorization": f"Bearer {api_key}"
+    }
+
+    payload = {
+        "model": "gpt-4-vision-preview",
+        "messages": [
+            {
+                "role": "user",
+                "content": [
+                    {
+                        "type": "text",
+                        "text": "What's in this image? Keep your response around 2-3 sentences."
+                    },
+                    {
+                        "type": "image_url",
+                        "image_url": {
+                            "url": f"data:image/jpeg;base64,{base64_image}"
+                        }
+                    }
+                ]
+            }
+        ],
+        "max_tokens": 300
+    }
+
+    response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)
+
+    return response.json()['choices'][0]['message']['content']
+
+def main():
+    img_response = identify_image("reference_img.jpg")
+    print(img_response)
+
+
+if __name__ == "__main__":
+    main()
