+++ openai-move-process.py
@@ -2,9 +2,7 @@ import os
 from dotenv import load_dotenv
 from openai import OpenAI
 from enum import Enum
-import base64
 import json
-import requests
 
 load_dotenv()
 
@@ -12,7 +10,6 @@ client = OpenAI(
     # This is the default and can be omitted
     api_key=os.environ.get("OPENAI_API_KEY"),
 )
-
 
 class Commands(Enum):
     MOVE_LEFT = "MOVE_LEFT"
@@ -86,69 +83,14 @@ def get_command_and_movement(text):
 
     return content_dict
 
-"""
-Identifies the object in an image using the OpenAI API's image recognition model.
-
-Parameters:
-- image_path (str): The file path of the image to be analyzed.
-
-Returns:
-- dict: A dictionary containing the identified object and its characteristics.
-"""
-def identify_image(image_path):
-    api_key = os.environ.get("OPENAI_API_KEY")
-
-    # Function to encode the image
-    def encode_image(image_path):
-        with open(image_path, "rb") as image_file:
-            return base64.b64encode(image_file.read()).decode('utf-8')
-
-    # Getting the base64 string
-    base64_image = encode_image(image_path)
-
-    headers = {
-        "Content-Type": "application/json",
-        "Authorization": f"Bearer {api_key}"
-    }
-
-    payload = {
-        "model": "gpt-4-vision-preview",
-        "messages": [
-            {
-                "role": "user",
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "What's in this image? Keep your response around 2-3 sentences."
-                    },
-                    {
-                        "type": "image_url",
-                        "image_url": {
-                            "url": f"data:image/jpeg;base64,{base64_image}"
-                        }
-                    }
-                ]
-            }
-        ],
-        "max_tokens": 300
-    }
-
-    response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)
-
-    return response.json()['choices'][0]['message']['content']
-
-
-
 def main():
     input_text = "Hey Spot, turn around twice"
-    #result = get_command_and_movement(input_text)
-    img_response = identify_image("reference_img.jpg")
-    # output = {
-    #     "movement": result["degrees"],
-    #     "command": Commands[result["command"]].value
-    # }
+    result = get_command_and_movement(input_text)
+    output = {
+        "movement": result["degrees"],
+        "command": Commands[result["command"]].value
+    }
     print(img_response)
-
 
 if __name__ == "__main__":
     main()
